Rename codegen::ndarray::naive::inner as ndarray::naive
